Display: Disable Dynamic Power Management of Screen
After 10 minutes the screen is automatically put in standby which is not managed
by the platform (especially the wake-up). This commit use the file /etc/issue
for disabling the Dynamic Power management of Screen using well known string
which is appended at build time. The issue file is used for any tty connection.