Boot: Automatic boot dependencies
On some borad we are not able to get the launcher executing when booting. The
problems comes from some missing dependencies.
gdp-hmi-controller need to be started after weston service
gdp-hmi-panel and background need to be started after controller service
gdp-hmi-launcher2 need to be started after controller, panel and background

some time we also observed that controller failed to start is why we set Restart
attribute to always. And it is important also to define the duration between two
start to 2 sec in order not to stress the controller and get unrecoverable
restart.