GDP: fix issue when booting without second SDcard
for persistence support we are using second sd-card. Unfortunately, in case the
second sd card is not present, the system is no more booting.

This commit fix problem introduced by commit
commit 2ce2a6feac3f4daad4410bffbc4253326c528958
    GDP: create persitence mounting point

Now before starting launcher we test if the sdcard is present or not and then
mount expected mounted directory only in the case it is present.