Add layer.conf and conf-notes.txt
Can't get contents of conf-notes.txt to be displayed even with setting
TEMPLATECONF.  But I think this is because the template directory is not used
here since it is noticed that the files already exist.  It is the case here
since the conf files have been created and committed to the repo already,
instead of re-created from a template every time.