Force gdp-hmi-panel to restart with the gdp-hmi-controller
This is necessary because the gdp-hmi-panel registers the PID of
the gdp-hmi-controller when it starts. If the gdp-hmi-controller
crashes, it restarts with a different PID, and is inaccessible to
the gdp-hmi-panel.

This patch fixes this by forcing the gdp-hmi-panel to be stopped when the
gdp-hmi-controller is (via gdp-hmi-panel Requires gdp-hmi-controller),
and also triggers the gdp-hmi-panel to start when the gdp-hmi-controller
is restarted (via gdp-hmi-controller Wants gdp-hmi-panel)

Signed-off-by: Jonathan Maw <jonathan.maw@codethink.co.uk>