Add a uinput-based virtual mouse to test GDP HMI
It consists of installing a script that uses python-uinput to click the
next button, the app icon, and the panel's "back" button on a loop.

Note that this requires a kernel with uinput enabled, which is not included
with this patch as it has been tested on the Porter board, which uses its
own kernel recipes.

Signed-off-by: Jonathan Maw <jonathan.maw@codethink.co.uk>