rvi-default-config: Add recipe
Add Yocto/OE recipe for providing default configuration files and
certificates for rvi_lib and Aktualizr. Although  this Yocto/OE recipe
is not mandatory it is very convenient in use cases when the user has
to configure Aktualizr (or alternative SOTA client) and to set device
ID based on the example provided by /etc/sota.toml.

[GDP-551] Integrate Aktualizr (SOTA Client C++)

Co-Authored-By: Phil Wise <phil@advancedtelematic.com>
Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>