chromium: Fix compilation for Raspberry Pi
The Raspberry Pi is identified as an ARMv7VE board and Yocto sets the
corresponding optimization flags. Unfortunately, Chromium sets its own
flags too, and assumes that any ARMv7 compilation is ARMv7A, which sets
a set of flags that conflict Yocto's.

Besides, meta-browser recipe for Chromium also expects the ARMv7A
architecture, and refuses to build if it's not set.

This patch adds certain overrides for ARMv7VE architectures, which let
the recipe behave like in the case of an ARMv7A build:

* Add armv7ve as a compatible architecture.
* Use armv7a FP tune flags.
* Add a set of source files that come from the armv7a build.
* Disable the incompatible armv7ve optimization flag.

[GDP-477] Integrate Chromium on a Wayland/Ozone stack

Signed-off-by: Jacobo Aragunde Pérez <jaragunde@igalia.com>