ci-build.sh: Store build info - hashes & env
Store a dump of submodule info and all environment variables in a file
inside the staging (artifacts) directory.

The reason is to save information about the build and store it together
with the artifacts so that they are easily correlated. Most obviously we
identify which source was built.  Go.CD provides some things like git
commit hash for the main repos in variables and many other environment
variables can also affect the build, so we dump all of them out as a
crude but effective way.

In addition, submodule hashes and status (e.g. dirty) are now logged and
any modifications to conf files are stored.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>