init.sh: Support more than one BSP layer per target
init.sh was already downloading only one particular BSP submodule for
the chosen target to avoid downloading BSP layers for targets that are
not built.  This change enables having more than one layer defined as
unique for a particular target.

The trigger for this was R-Car gen3, which added two layers, only used
(so far) by that hardware target, namely meta-linaro and
renesas-rcar-gen3.  Both of those are now designated as BSP layers for
this hardware target. (Before this change, meta-linaro was unnecessarily
cloned for all other targets too)

As an implementation detail, qemux86-64 which needs no BSP layer was
added to BSP array with an empty value, which avoids the special logic
that was in place before.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>