init.sh: Rename variable $machine to $target
Previously, the hardware target passed to init.sh was identical with the
$MACHINE variable used in the layers.  As of late, the name of the
embedded board is sometimes used instead. From now on, init.sh shall be
passed a designation of the hardware target, i.e. typically the embedded
board name. The variable name was changed to reflect that and reduce
confusion.  This resolves the ref. ticket (although in a different way).

Tweaked usage printout.

[GDP-572] Align R-car target name with $MACHINE

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>