chromium: gcc fix #66758 to meta-linaro-gdp sublayer (2)
The .bbappend for gcc-source_linaro introduced in the previous commit
would break the build for other targets since there is no
gcc-source_linaro.bb to append to for those builds. The append must be
applied only for relevant targets.

This is a quick fix. We expect the patch to be in the meta-linaro
toolchain eventually and we can then upgrade and drop this patch and
layer.  In the meantime we can carry it here.

A new sublayer meta-linaro-gdp was added to keep the fix in the same way
other BSP fixes are done, and the sublayer is intended to be
(temporarily!) included only in the conf for the R-Car build.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>

EDIT: During m-g-d subtree add work, I found that this commit in its
then current form (09983e4) was lacking the configuration file change
that enables the sublayer. That ought to be part of the same commit.
That conf file change was intermixed with a submodule update in
3f1e5e1a.  The submodule no longer exists, so cherry-picking that makes
no sense.  Since we are rewriting this branch anyway this commit is
amended to include the conf file change.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>