chromium: Include multi-seat implementation
Modify recipe repositories via .bbappend to point to the fork and
branch of ozone-wayland containing the multi-seat implementation.
It was necessary to remove the original repo from SRC_URI and append
the new one.

The patch 0006-Remove-GBM-support-from-wayland.gyp.patch that comes
with upstream meta-browser was removed from the SRC_URI because it
didn't apply on top of the multi-seat branch. It was properly
rebased and committed into that branch.

[GDP-604] Finalize and document Chromium Multi-Seat feature

Signed-off-by: Jacobo Aragunde Pérez <jaragunde@igalia.com>