git-submodule-bump: Always edit commit after
Always view/edit the git-submodule-bump commit after creating it.

I find that 99% of the time I want to check that the commit message is
alright, and often also add additional comments.  Opening the editor
during the commit operation also enables us, as usual, to abort the
commit by saving an empty message.

Minor: If we forget to quote the commit summary on the command line, it
will now work (most of the time) to still get all the words included.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>