meta-ivi: Use wayland-ivi-extension v1.11.1
Bumped meta-ivi to latest on 12.0 branch. This also means we can remove
the locally applied patch for a crashing bug which is fixed in in the
newer version of wayland-ivi-extension.  At the moment that patch is
applied on chromium feature branch only, so there's nothing (yet) to
remove on master branch, but I still want this commit on master.  The
patch removal will thus follow this in a separate commit.  I won't
squash them since it would just complicate merging the feature branch
later.

b0adb53 wayland-ivi-extension: update to release v1.11.1

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>