meta-genivi-dev: Move .bbappends into layername dirs
All .bbappend that were in meta-genivi-dev layer (with one or two
remaining exceptions that are noted as tickets for the future) are moved
to fully qualified subdirectory names indicating which layer they are
extending/modifying.

This is according to previous proposal/discussion (GDP-609, which links
to proposals, mailing-list discussion and the documented naming
convention resulting from rough consensus).

It may not be perfect, but it was a rough consensus and something we
want to try for a while.

[GDP-609] Create a strategy for where to put stuff in the...

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>