Move .bbappends into "namespace" directories
All .bbappend (with one or two remaining exceptions) are moved to
fully qualified subdirectory names indicating which layer they are
extending/modifying.

This is according to previous analysis (GDP-609, which links to
discussion material proposals, mailing-list discussion and the
documented naming convention resulting from the consensus).

It may not be perfect, but it was the rough consensus and something we
want to try for a while.

[GDP-609] Create a strategy for where to put stuff in the...

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>