meta-genivi-dev: Remove more patches, leftover from merge
These patches do not seem to be applied by any recipe any longer.  I
would guess these were removed in pyro upgrade and the result of a
faulty merge.  I'm being brief because it is unlikely we merge this
branch/PR anyway - I'm now just bringing it up to the right status.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>