init.sh: Add a simple help to avoid locale errors
Bitbake (python in fact) requires a terminal locale that supports UTF-8.
Most of the time users will have that but not always.  Also in some
simplified environments like CI-builds it might be missing.  For
simplicity's sake I figured we can have init.sh set a locale like
English/US/UTF-8 in case it needs to be done.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>