Revert "init.sh: Add a simple help to avoid locale errors"
This reverts commit 45537cc40e8fc9fc2608beb150535f3e220273f2.

This is an immediate hotfix since it was a bad merge.

First analysis (hypothesis) of problem:

Because ci-build runs with -e flag, and the pipeline used in setLocale()
function is set up (deliberately) to fail when no appropriate locale
is found, then script will exit and compilation will halt immediately.

The PR was successfully reviewed but we accidentally missed that the
pull request had not been build-checked for some reason.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>