meta-erlang: Pin OTP package to 18.2.3
Signed-off-by: Oscar Andreasson <oscar.andreasson@pelagicore.com>

We are including the MD5 tarball checksum mismatch issue (see
meta-erlang:b6375e7).  Those patches were merged to the tip of upstream
master for meta-erlang.  Instead of backporting we bumped the layer up
to this latest version.  We found incompatibilities with the version of
OTP (19.0) that is then built by default, so for the moment we pin the
version to the previously known one.

I split the fix from the layer update and dropped the layer fork (back
on upstream again).  I also corrected the commit summary accordingly.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>