nativesdk-dbus: Don't install target files in SDK
There is a bbappend to dbus that installs a few extra systemd service
files and things.

Compilation failed during SDK build because these installed files were
not shipped in a package (QA check).  By using a class-override
statement, do_install_append_class-target, we ensure the files are
installed only for the target, not for the SDK build.

It is also being discussed if this append should be there at all,
or if there's a better solution: (GDP-705).

[GDP-713] https://at.projects.genivi.org/jira/browse/GDP-713

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>