nativesdk-qemu: remove pulseaudio dependency
We discussed disabling the build of qemu completely since users use the
one installed by their distribution, so the configuration options do not
matter.  I was unable to remove it fully for nativesdk build and the
compilation bug must be fixed, so for this workaround I removed the
compilation issue instead.  A similar change was discussed also for
meta-ivi (see related tickets).

[GDP-713] - SDK fails to build

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>