Add script to update meta-ivi's dependencies
When meta-ivi updates, it defines the compatible commits/versions of
dependent layers such as poky and meta-openembedded in its README.

After updating meta-ivi, running this script will update the layers in
GDP to the versions proposed by the baseline.  Sometimes we may choose
to deviate from it, but it is a starting point that is likely to be
compatible.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>