ci-build: Restore user.name/email
The ci-build script would clobber the user's user.name and user.email
settings in git, because that makes no difference in a throw-away CI
environment.  It is required (in some cases, such as when git is used as
patch tool) that there is SOME setting for user/email in the CI build.
But since ci-build.sh could be useful to run also locally, the original
settings must be restored when the script ends.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>