ci-build.sh: Add GENIVI MIRROR + PREMIRROR option
Adding environment variables to define MIRRORS and PREMIRRORS variables.
Only one URL for each is supported here.  PREMIRROR is by default empty,
and MIRROR is by default set to a GENIVI hosted mirror.  (MIRRORS is
used by Yocto as a fallback if all other source fetch options failed)

For PREMIRRORS it is normally more convenient to use the own-mirrors BB
class, but since MIRRORS need to be explicitly defined, I used the same
format for PREMIRRORS.

It's quite arbitrary but I chose to apply the PREMIRROR definition first
(prepend) and MIRROR definition last (append), in case there were
any settings already.  Seems somewhat logical that this is the desired
behavior.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>