WIP -- try blacklist iotivity from hardening flag.
This seems more a -Werror type or problem, but
just to check

WIP!  Adding to this commit...:

old commit info:

Blacklist some packages from compiler hardening flags

The gcc "hardening" flags were applied in meta-ivi distro. It was
requested by Phong Tran in PR #98 for GDP, but the flags are applied to
the whole distro, which is why they are placed in the meta-ivi distro
definition instead.

Still, a few packages are reported to fail to build with these flags so
they cannot be applied, or they can be partially applied.  This commit
adjusts flags for those packages that are are in GDP only, not meta-ivi.

[GDP-754] Compiler Hardening Flags

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>