AudioManager: Port bbappend to AM 7.6
This is a simple rename of the version number on the bbappend
to match the main recipe in meta-ivi.

A quick look suggests that the patch should apply to the new tag without
change.  Unless the build configuration flags have changed somehow
then those parts should be fine too, and of course the systemd file
should need no change, I'm guessing.

In any case, there is more to fix on the AM integration and this
is only to make it compile sucessfully with the latest meta-ivi.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>