packagegroup-gdp-dev: add udisks2 packages
udisks2 provides a D-Bus enabled daemon to manage and manipulate
storage devices and a command-line tool "udisksctl" to control
the daemon.

udisks2-disk-manager is a service that we have created to automount
storage devices on hot-plug and shall only be seen as an example at
best. A "DeviceManager" should talk directly to the udisks2 daemon
over D-Bus and also be responsible of setting up policies in regard
to security, which is completly bypassed in this case.

udisks2-disk-manager will mount devices to the following path:

    /run/media/root/<LABEL>

[GDP-252] Add support for udisks2 to the GDP image

Signed-off-by: Mirza Krak <mirza.krak@endian.se>