Remove qtwebkit from SDK
It does not compile with gcc7 and at the moment we can not
prioritize backporting fixes for this. Solution is to remove
it from the SDK as we currently do not have any users of it and
we do provide a Chromium build by default which should be used
instead.

Signed-off-by: Mirza Krak <mirza.krak@endian.se>

Modified the commit after discussion to use a bbappend on
the qt5 toolchain package group instead of variables est
in the genivi-dev-platform-sdk.bb image definition.
Small cleanup of unncecessary variable remains in the
...sdk.bb

[GDP-788] Fix SDK build failing on qtwebkit (GDP 14/rocko)

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>