systemd: Configure without polkit (for now)
udisks2 depends on polkit which installs the same catalog as the
systemd.bb does if PACKAGECONFIG[polkit] is enabled.  There seems to be
two options: - keep polkit enabled for systemd but redefine the
installation step to fix the conflict for the directory installation, or
remove polkit feature in systemd build.

Disabling the feature will also configure systemd build with
--disable-polkit.

Polkit can be used to control which systemd services can be controlled
by regular users.  I don't think that feature is required here and it
seems more tailored for a multi-user "desktop" situation.  Assuming that
we don't use polkit with systemd now, the feature can be avoided for the
moment.

If it's ever needed again, then the other solution can be investigated.

[GDP-791] Transaction check error polkit conflicting with systemd

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>