templates: renesas & local.inc: remove package_rpm
RPM was our primary package manager which means that the
do_rootfs would use dnf to install all the required packages.

Currently RPM has performance issues when running inside a
docker container, which our builds normally do on our build
servers but it is also quite common to run this on developer
machines as well.

The performance issues are rapported and fixed in upstream RPM,
https://bugzilla.redhat.com/show_bug.cgi?id=1537564.

But since we do not rely on using RPM we simply remove it instead
of trying to backport and maintain a patchset for a couple of
Yocto release. Currently only poky/master has the "fixed" RPM
package.

[GDP-793] rpm in docker is slowing down builds

Signed-off-by: Mirza Krak <mirza.krak@endian.se>

Amended: Removed also from renesas-rcar-gen3.local.inc because it's
duplication and @slawr indicated there should be no special need.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>