connman: overwrite /var/lib/connman/wired.config
Instead of editing it. This resolves an issue when running a libostree
image where /var is the persistant area and is not populated by default.
Meaning that even thoug the /var/lib/connman/wired.config is installed
during build time it will not be present when the system is booted.

[GDP-794] Add support to boot and test libostree images in CI loop

Signed-off-by: Mirza Krak <mirza.krak@endian.se>