meta-genivi-dev: oeqa: add devpackages test
A sanity test making sure that no development packages are installed
on target unless "dev-pkg" is set in IMAGE_FEATURES.

It is quite easy to add a dependecy somewhere on a -dev package which
would pull in libc headers and all the dev packages that that specific
package depends on without any warnings or errors. This test is a
way of detecting this and error out if it happens.

[GDP-787] - Remove development (.h, .pc) files from target image

Signed-off-by: Mirza Krak <mirza.krak@endian.se>