automotive-message-broker: remove rdep on ${PN}-dev
The shlibdeps step in do_package parses libraries and executables
to find rdeps and automatically adds them to package REPENDS if
a package is found that provides that library.

One problem with above, one of the executables links with
libamb-plugins-common.so which is part of the
automotive-message-broker-dev package and this is turn meant that
it pulled in development files on target for this recipe and all its
dependencies.

To fix this one most set PRIVATE_LIBS to mark .so files that are for
internal usage only, which libamb-plugins-common.so is in this case.

[GDP-787] - Remove development (.h, .pc) files from target image

Signed-off-by: Mirza Krak <mirza.krak@endian.se>