positioning: fully utilize the cmake.bbclass
The recipe overrided do_configure and do_compile with "custom"
implementations, but they do not add anything specific that the
cmake.bbclass does not already support. So drop the custom
overrides.

Needed to adjust some paths since ${B} and ${S} where the same
before, which they are not anymore due to cmake.bbclass making
the seperation.

Also added INSANE_SKIP_* on the actual QA errors instead of
disabling QA completly.

Signed-off-by: Mirza Krak <mirza.krak@endian.se>