packagegroup-gdp-dev: add avahi-daemon
We have a silent dep on avahi-daemon for our acceptance tests.
https://git.yoctoproject.org/cgit/cgit.cgi/poky/tree/meta/lib/oeqa/runtime/cases/systemd.py#n102

When we removed -dev packages from target we also removed something
that depended on avahi-daemon because it was installed before.

By adding this in our -dev packagegroup we now have an explicit
dependency instead of relying on pure luck.

[GDP-787] - Remove development (.h, .pc) files from target image

Signed-off-by: Mirza Krak <mirza.krak@endian.se>