image: Add compilers, and dependencies for VCIVING
For a development version of the image with on-target compilers
(and some other useful tools) + various -dev packages needed for
the VCIVING dependencies.

Also -dev packages for python2/3 and perl, and pip for easy installation
of python dependencies on target.

This also means the test that fails if any dev-packages end up on target
is removed here.

(WIP)