minifi-cpp: initial recipe for Apache MiNiFi
Proof of Concept state and have not tested enabling all the extensions,
but the default and core components build and deploy properly.

There is systemd service (minifi.service) that starts MiNiFi on
boot but it does not collect any data at the moment and a proper
configuration file must be crafted for this.

[GDP-801] - Write recipe for Apache NiFi MiNiFi
[GDP-805] - Create systemd service for minifi-cpp

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>