genivi-dev-platform: install connman-conf for QEMU
connman-conf is a set of scripts and services that allow you
to set the IP address of the QEMU using Linux kernel arguments
("ip="). This is used by our acceptance tests.

connman-conf is set as RRECOMMENDS in the connman recipe, but
we actually require this otherwise our tests will fail.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>