Remove iotivity & nodejs (but not the layers*)
At the moment the meta-layers for iotivity are quite out of date
and we don't have the bandwidth to update them.  There are also some
immediate compilation issues right now that is holding back progress on
other fronts so we have to consider dropping it from master.

I've had some contact with previous and new maintainers, but it is
still lagging behind.  We can always reverse this decision and it is
possible it will remain on a certain branch/variant of GDP or that we
get some sort of update later on.

This also removes a dependency on nodejs

* This patch removes the layers from bblayers.conf, but does not not
modify .gitmodules at this time because of some issues in CI when the
same pipeline switches between branches where the submodule is included
vs removed.  If this commit gets merged to master and, once ALL branches
are up to date with that, we can proceed to remove the submodule
properly to avoid these issues.

[GDP-809] Updating or removing iotivity (for the moment)

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>