genivi-dev-platform: set custom IMAGE_FSTYPES
It seems to be an convention for each machine to set IMAGE_FSTYPES
differently and it is pretty common that it set using "_append" override.
This means that we want to modify which IMAGE_FSTYPES we must use the
"_remove" override if we are not interested in the images set by machine.

Using "_remove" is not future proof, as new image types might be added
meaning that our "_remove" will not take care of it.

IMAGE_FSTYPES can be overridden by an image recipe, and this is what
this commit is trying to utilize. By defining CUSTOM_IMAGE_FSTYPE, which
defaults to "tar.gz". Meaning that if we it is not set only tar.gz will be
generated, and we need a default otherwise it will error out.

The supported boards have been updated utilizing "wic.gz" where it is
supported as this is the convection now days in Yocto.

WIC images can be written directly to an SD/eMMC, example:

    zcat genivi-dev-platform-m3ulcb.wic.gz | sudo dd of=/dev/mmcblk0

[GDP-810] - IMAGE_FSTYPES cleanup

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>