Remove support for Renesas R-Car, Gen 2
As decided by GENIVI/GDP/baseline discussions long ago, boards for
R-Car "generation 2" (e.g. koelsch, porter, silk) will not be supported
going forward, simply because of the tradeoff between interest and
effort on maintenance.

These targets have been left in as available for quite a while but only
as community-supported.  They are now removed completely from the
build/init scripts.

There is nothing that prevents a community member to pick up the thread
and reintroduce support of course.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>