ci-build.sh:  Remove unnecessary def of $MACHINE
Each of the machine-specific local.conf that is specified (by init.sh)
includes an explicit assignment of MACHINE.  It should not be necessary
to have it as environment variable also (as far as I can tell).  It
seems to be a legacy detail that has been handled incorrectly during
a transition from using $MACHINE to $TARGET as the input variable to the
scripts.

+ removed two superfluous lines

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>