Correct LAYERDEPENDS for append-layers in meta-genivi-dev
Many of the append-layers were not correctly depending on the layers
they are modifying.  This is corrected, first of all.

We also debated whether these should continue to depend on the main
meta-genivi-dev layer or not.

Technically speaking, for many/most of the fixes it probably does not
require _also_ something from meta-genivi-dev. In other words, they
should probably apply cleanly on top of the standard layers they modify
without problem.  This suggests to not add any additional LAYERDEPENDS.

On the other hand, these dependencies made some sense to communicate
that the appends are specific for genivi-dev-platform and _intended_ to
be specific (because we try to push out to upstream and other as much as
possible and remove the local change).  It also seems very unlikely to
reuse these layers as they are in other contexts.

But, in the end such communication seems to be a misuse of the variable,
and there seems to be little purpose to putting a hard restriction in
here - so it was removed.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>