Correct LAYERDEPENDS for append-layers in meta-genivi-dev
Many of the sub-layers (append-only layers) were not correctly depending
on the layers they are modifying.  This is corrected, first of all.

We also debated whether these should continue to depend on the main
(meta-)genivi-dev layer or not.

Most often we expect there is no technical dependency, so these appends
should apply cleanly on top of the layers they modify without needing
genivi-dev also. This suggests to not keep genivi-dev as LAYERDEPENDS.

On the other hand, depending on genivi-dev made some sense to
communicate that the appends are specific for genivi-dev-platform and
_intended_ to be specific (because we try to push out to upstream and
other layers as much as possible, so only unique things are left,
at least ideally).  That makes it less likely to reuse these minor
adaption layers in other contexts without modification.

But, in the end simply communication seems to be a misuse of the
variable, and we decided there is little point in adding a hard
restriction through dependency here -- so dependency on genivi-dev was
removed.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>