meta-renesas (append): Correct the sublayer name/path
A minor quality fix that adjusts the path of the append-layer for
meta-renesas, since the layer it applies to has a subdirectory/ sublayer
named meta-rcar-gen3.

This is just to strictly follow the naming standard of append-layers in
GDP which requires the path in the append-layer to be identical to the
path it is modifying in the main layer.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>