Remove meta-gnome layer
Looks like chromium dependend on gnome keyring but after the switch to
meta-lgsvl-browser layer I can't find this, or any dependency any longer.
Let's see if it builds without meta-gnome.

The trigger for this was that meta-gnome started requiring that x11 be
enabled (warning only), but this system (including chromium) is of
course wayland-based.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>