ci-build: Allow override some parts of STANDARD_RELEASE_BUILD
STANDARD_RELEASE_BUILD=true is a shortcut to set several other settings
at once, for example COPY_LIC_DIRS=true, but it should still be possible
to override those individual settings with a different value.  There was
previously a comment to warn about values being forced, because what is
now implemented is more the intended behavior.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>