Revert "meta-ivi: + dep. layers - update to thud"
This reverts commit a4b8247a8474ffebdca97c70082284c19b7d6444.

Sorry, this have been merged or pushed by mistake.  The work is on a
separate thud branch and it needs some more commits than just this layer
update.  Also the aim is to clean up that branch a bit before merging to
master.

Since it's been published a while, an explicit revert seemed in order.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>